# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API - Swagger Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login an User account",
                "operationId": "user-login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "text"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Validation Failed"
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Logout an User account",
                "operationId": "user-logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Validation Failed"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/v1/user/create": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create a User account",
                "operationId": "user-create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "address",
                                    "phone_number"
                                ],
                                "properties": {
                                    "first_name": {
                                        "description": "User firstname",
                                        "type": "text"
                                    },
                                    "last_name": {
                                        "description": "User lastname",
                                        "type": "text"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "text"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "User password",
                                        "type": "password"
                                    },
                                    "address": {
                                        "description": "User main address",
                                        "type": "text"
                                    },
                                    "phone_number": {
                                        "description": "User main phone number",
                                        "type": "text"
                                    },
                                    "is_marketing": {
                                        "description": "User marketing preferences",
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Page not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "422": {
                        "description": "Validation Failed"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}